feat: MHX RS4xx interface (closes uavos/apx-releases#33)
To connect `mhx` serial interface `RS4xx` to a PC, it must be configured as `MANDALA` or `DOWNLINK` serial port. Moreover, when the `mhx` function is selected as `gcu` - the RS232 port can not be used as protocols driven virtual serial port, because it always acts as bus interface for node locking protection. `RS232` port, when it is used to connect to a PC is always have default baud rate **460800**.